requirements.txt
Django==4.0.4
Pillow==9.1.0
python-dotenv==0.20.0
whitenoise==6.2.0
psycopg2==2.9.3
boto3==1.24.48
django_storages==1.13.1
djangorestframework==3.13.1
